+++ src/components/tiles/advanced/AiAnalysis.tsx
@@ -23,7 +23,7 @@ const AiAnalysis = () => {
     {
       x: x1,
       type: 'histogram',
-      name: 'Fire Detected',
+      name: 'Feuer',
       opacity: 0.7,
       marker: {
         color: 'red',
@@ -32,7 +32,7 @@ const AiAnalysis = () => {
     {
       x: x2,
       type: 'histogram',
-      name: 'Smoke Detected',
+      name: 'Rauch',
       opacity: 0.5,
       marker: {
         color: 'grey',
@@ -58,7 +58,7 @@ const AiAnalysis = () => {
               autosize: true,
             }}
 
-            style={{ width: '100%', height: '100%' }}
+            style={{ width: '100%', height: '90%' }}
             config={{ responsive: true }}
           />
